:bug: avoid deleting default file :memo: add comment

+++ router/v1/upload_thumbnail.go
@@ -32,6 +32,7 @@ func (h *Handler) GetClubThumbnail() gin.HandlerFunc {
 	}
 }
 
+// XXX: unnecessary func
 func (h *Handler) UploadClubThumbnail() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		file, err := ctx.FormFile("file")
@@ -115,6 +116,10 @@ func (h *Handler) checkThumbnail(file *multipart.FileHeader) error {
 }
 
 func (h *Handler) deleteSavedThumbnail(file string) error {
+	if file == "thumbnails/default.png" {
+		return nil
+	}
+
 	if err := os.Remove(file); err != nil {
 		h.logger.Error(err.Error())
 		return err
